% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-license.R
\name{check_raw_lines}
\alias{check_raw_lines}
\title{Compare count of lines in raw file to imported data frame}
\usage{
check_raw_lines(df, raw_file)
}
\arguments{
\item{df}{data frame into which raw_file was loaded}

\item{raw_file}{path to raw data file}
}
\description{
Generally speaking the raw file will have one more line (header) than the 
data frame has rows. This function is mostly a wrapper for
\code{\link[salic]{count_lines_textfile}}
}
\seealso{
Other functions for preparing license data: 
\code{\link{date_to_char}()}
}
\concept{functions for preparing license data}
