% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_demo}
\alias{cust_dup_demo}
\title{Demographic breakouts for customer duplication}
\usage{
cust_dup_demo(cust, cust_dup, demos = c("sex", "cust_res", "birth_year"))
}
\arguments{
\item{cust}{input customer table}

\item{cust_dup}{table of customers (with dups) produced by 
\code{\link{cust_dup_pull}}}

\item{demos}{name(s) of demographic variables to check}
}
\description{
Runs \code{\link{cust_dup_pct}} across demographic variables. By default
requires sex, cust_res, and birth_year to be present in the customer tables.
Returns a data frame with breakouts for each specified demographic.
}
\seealso{
Other finalize production data: 
\code{\link{cust_dup_demo_plot}()},
\code{\link{cust_dup_identify}()},
\code{\link{cust_dup_pct}()},
\code{\link{cust_dup_pull}()},
\code{\link{cust_dup_year}()},
\code{\link{res_id}}
}
\concept{finalize production data}
