% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res-identify.R
\name{cust_dup_demo_plot}
\alias{cust_dup_demo_plot}
\title{Plot results of cust_dup_demo()}
\usage{
cust_dup_demo_plot(df_pct)
}
\arguments{
\item{df_pct}{output of \code{\link{cust_dup_demo}}}
}
\description{
Plot results of cust_dup_demo()
}
\seealso{
Other finalize production data: \code{\link{cust_dup_demo}},
  \code{\link{cust_dup_identify}},
  \code{\link{cust_dup_pct}}, \code{\link{cust_dup_pull}},
  \code{\link{cust_dup_year}}, \code{\link{res_id}}
}
\concept{finalize production data}
