% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_nomissing}
\alias{cust_dup_nomissing}
\title{Set cust_id to cust_id_raw where there are missing values in select variables}
\usage{
cust_dup_nomissing(cust, vars)
}
\arguments{
\item{cust}{input customer table}

\item{vars}{variables to look for missing values}
}
\description{
This prevents a situation in which multiple customers with missing values (e.g.,
in names) are identified as the same customer. To be run following
\code{\link{cust_dup_identify}}
}
\seealso{
Other finalize production data: 
\code{\link{cust_dup_demo_plot}()},
\code{\link{cust_dup_demo}()},
\code{\link{cust_dup_identify}()},
\code{\link{cust_dup_pct}()},
\code{\link{cust_dup_pull}()},
\code{\link{cust_dup_year}()},
\code{\link{res_id}}
}
\concept{finalize production data}
