% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_pct}
\alias{cust_dup_pct}
\title{Summarize percent of customers with extra rows in raw table}
\usage{
cust_dup_pct(cust, cust_dup, return_msg = TRUE)
}
\arguments{
\item{cust}{input customer table}

\item{cust_dup}{table of customers (with dups) produced by 
\code{\link{cust_dup_pull}}}

\item{return_msg}{If TRUE, returns a message with the duplication percentage,
otherwise, the duplication percentage measure}
}
\description{
Percent of customers: (based on deduplicated count) with extra
rows in the raw customer ID table. This will give a lower percentage
than \code{\link[salic]{check_dups}} because it only counts duplicated
customers once. I consider this to be a more meaningful measure of the 
customer duplication rate.
}
\seealso{
Other finalize production data: 
\code{\link{cust_dup_demo_plot}()},
\code{\link{cust_dup_demo}()},
\code{\link{cust_dup_identify}()},
\code{\link{cust_dup_nomissing}()},
\code{\link{cust_dup_pull}()},
\code{\link{cust_dup_samp}()},
\code{\link{cust_dup_year}()},
\code{\link{res_id}}
}
\concept{finalize production data}
