% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_pull}
\alias{cust_dup_pull}
\title{Pull customer records where duplicates were found}
\usage{
cust_dup_pull(cust)
}
\arguments{
\item{cust}{input customer table}
}
\description{
This table is intended to be stored in case deduplcation needs to be checked
downstream. The output only includes those customers in which a duplicate was 
found, containing all corresponding rows from the raw customer table.
}
\seealso{
Other finalize production data: \code{\link{cust_dup_demo_plot}},
  \code{\link{cust_dup_demo}},
  \code{\link{cust_dup_identify}},
  \code{\link{cust_dup_pct}}, \code{\link{cust_dup_year}},
  \code{\link{res_id}}
}
\concept{finalize production data}
