% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_samp}
\alias{cust_dup_samp}
\title{Pull a sample of duplicated customers and their sales}
\usage{
cust_dup_samp(
  cust_dup,
  sale,
  samp_size = 10,
  vars = c("cust_id", "cust_id_raw", "year", "first", "last", "dob", "zip4dp")
)
}
\arguments{
\item{cust_dup}{table of customers (with dups) produced by 
\code{\link{cust_dup_pull}}}

\item{sale}{data frame with uncorrected customer IDs}

\item{samp_size}{number of customers to pull in sample}

\item{vars}{variables to include in output}
}
\description{
Only select variables are included, with a \code{\link[dplyr]{distinct}}
operation run at the end.
}
\seealso{
Other finalize production data: 
\code{\link{cust_dup_demo_plot}()},
\code{\link{cust_dup_demo}()},
\code{\link{cust_dup_identify}()},
\code{\link{cust_dup_nomissing}()},
\code{\link{cust_dup_pct}()},
\code{\link{cust_dup_pull}()},
\code{\link{cust_dup_year}()},
\code{\link{res_id}}
}
\concept{finalize production data}
