% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{cust_dup_year}
\alias{cust_dup_year}
\title{Annual dynamics for customer duplication}
\usage{
cust_dup_year(cust, cust_dup, sale)
}
\arguments{
\item{cust}{input customer table}

\item{cust_dup}{table of customers (with dups) produced by 
\code{\link{cust_dup_pull}}}

\item{sale}{sale table for exploring year-to-year duplication patterns}
}
\description{
License system changes often lead to a spike in customer dups in comparing
new and old data. It can be useful to be aware of this.
}
\details{
Note that this function might not reveal a break point (i.e., from a change
to a new licensing system), the duplication identification would probably 
need to be iterative, so will probably need to rewrite this function to reveal 
such a pattern. This wasn't a priority in Dec 2019.
}
\seealso{
Other finalize production data: 
\code{\link{cust_dup_demo_plot}()},
\code{\link{cust_dup_demo}()},
\code{\link{cust_dup_identify}()},
\code{\link{cust_dup_nomissing}()},
\code{\link{cust_dup_pct}()},
\code{\link{cust_dup_pull}()},
\code{\link{res_id}}
}
\concept{finalize production data}
