% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{data_check_standard}
\alias{data_check_standard}
\title{Check standard data (cust, sale) formatting rules}
\usage{
data_check_standard(cust, lic, sale, sale_year1 = 2009)
}
\arguments{
\item{cust}{data frame: customer table (primary key = "cust_id")}

\item{lic}{data frame: license types table (primary key = "lic_id")}

\item{sale}{data frame: transactions table (foreign keys = "lic_id", "cust_id")}

\item{sale_year1}{First allowed year for sale dates}
}
\description{
Produces warnings if any checks fail (stays silent on success).
This function is simply a wrapper for several calls to \code{\link[salic]{data_check_table}} 
and \code{\link[salic]{data_foreign_key}}. Rules are designed to ensure:
\itemize{
  \item primary keys are unique and non-missing
  \item all required variables are present
  \item variables only contain prescribed values
  \item foreign keys are present in relevant primary key table (e.g., all
  sale$cust_id can be found in cust$cust_id)
}
}
\seealso{
Other functions to check data format: 
\code{\link{variable_allowed_dates}()}
}
\concept{functions to check data format}
