% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-license.R
\name{date_to_char}
\alias{date_to_char}
\title{Convert all date columns in a data frame to character}
\usage{
date_to_char(df)
}
\arguments{
\item{df}{input data frame}
}
\description{
A convenience function to prepare for sqlite, which doesn't translate R dates 
when writing data. It's a very slow operation, so not efficient with regard to
computation, but the alternative is to specify as character on load; 
a tedious thing to do with a larger number of variables.
}
\concept{functions for preparing license data}
