% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-license.R
\name{date_truncate}
\alias{date_truncate}
\title{Set unreasonable transaction date values to missing}
\usage{
date_truncate(df, date_var = "dot", year_var = "year")
}
\arguments{
\item{df}{frame that holds date variable}

\item{date_var}{name of date variable}

\item{year_var}{name of year variable}
}
\description{
This function asssumes that only values within the range covered by the "year"
variable are reasonable (i.e., year and dot should correspond)
}
\examples{
# sale <- date_truncate(sale)
}
\seealso{
Other functions for preparing license data: 
\code{\link{check_raw_lines}()},
\code{\link{date_to_char}()},
\code{\link{recode_date}()}
}
\concept{functions for preparing license data}
