% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-license.R
\name{plot_dates}
\alias{plot_dates}
\title{Plot sale date distributions}
\usage{
plot_dates(sale, var = "dot", angle = 30)
}
\arguments{
\item{sale}{sales data frame with year & var to plot}

\item{var}{date variable name to count}

\item{angle}{axis.text.x angle for plot}
}
\description{
Convenience function to check for unusual patterns or variation between years
}
\concept{functions for validating license data}
