% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_lic_types}
\alias{plot_lic_types}
\title{Make a heatmap of license type customers by year}
\usage{
plot_lic_types(df, truncate = 30)
}
\arguments{
\item{df}{Count of license types as returned by \code{\link{summary_lic_types}}}

\item{truncate}{Truncation point for "description" variable. If NULL, no
truncation will be performed.}
}
\description{
Make a heatmap of license type customers by year
}
\examples{
# sale <- left_join(sale, lic, by = "lic_id")
# df <- summary_lic_types(sale)
# plot_lic_types(df)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()},
\code{\link{summary_date}()},
\code{\link{summary_duration}()},
\code{\link{summary_lic_types}()},
\code{\link{summary_year_dot}()}
}
\concept{functions to summarize license data}
