% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-license.R
\name{recode_date}
\alias{recode_date}
\title{Convert date variables from character to date format}
\usage{
recode_date(df, date_var, fun)
}
\arguments{
\item{df}{data frame that contains date variable}

\item{date_var}{date variable stored as character}

\item{fun}{function used to convert from character to date format}
}
\description{
A convenience function that also includes a summary of any input values
thate weren't converted to dates (i.e., missing in output).
}
\examples{
# recode_date(cust, "dob", function(x) lubridate::ymd(str_sub(x, end = 10)))
}
\seealso{
Other functions for preparing license data: 
\code{\link{check_raw_lines}()},
\code{\link{date_to_char}()},
\code{\link{date_truncate}()}
}
\concept{functions for preparing license data}
