% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res-identify.R
\name{res_id}
\alias{res_id}
\alias{res_id_type}
\alias{res_id_other}
\alias{res_id_cust}
\alias{res_id_summary}
\title{Identify sales residency}
\usage{
res_id_type(sale, lic, var = "lic_res",
  caption = "Identification: by license type")

res_id_other(sale,
  caption = "Identification: other sales by the same customer")

res_id_cust(sale, cust, var = "cust_res",
  caption = "Identification: by customer address")

res_id_summary(sale, caption = "")
}
\arguments{
\item{sale, cust, lic}{license data frames}

\item{var}{variable name that holds reference residency variable}

\item{caption}{text to be printed above identification summary}
}
\description{
The intended use is to prioritize residency identification using 3 different
methods: (1) by license type, (2) by other sales by the same customer, and 
(3) based on customer address.
}
\section{Functions}{
\itemize{
\item \code{res_id_type}: Identify sales residency based on license type

This is the clearest identifier of residency, so it should be given first 
priority in identifying residency at the transaction level.

\item \code{res_id_other}: Identify residency based on other sales by customer

Based on the assumption that recent sales by the same customer can be used
to identify residency. More recent sales are given priority over distant sales.

\item \code{res_id_cust}: Identify residency based on customer address

\item \code{res_id_summary}: Summarize current residency percentages

This is intended to be run as part of the residency identification functions
}}

\seealso{
Other finalize production data: \code{\link{cust_dup_demo_plot}},
  \code{\link{cust_dup_demo}},
  \code{\link{cust_dup_identify}},
  \code{\link{cust_dup_pct}}, \code{\link{cust_dup_pull}},
  \code{\link{cust_dup_year}}
}
\concept{finalize production data}
