% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{setup_template}
\alias{setup_template}
\title{Setup template scripts for preparing agency data}
\usage{
setup_template(folder = "code/1-prep-license")
}
\arguments{
\item{folder}{location for template scripts}
}
\description{
The expected use assumes that the user has already setup a project directory
(e.g., using Rstudio). This function populates a subfolder with template
scripts.
}
