% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_date}
\alias{summary_date}
\title{Summarize date distribution by year}
\usage{
summary_date(df, date_var, yrs, samp_size = 10000)
}
\arguments{
\item{df}{data frame with date variable}

\item{date_var}{name of date variable}

\item{yrs}{years of interest}

\item{samp_size}{sample size to use for efficiency}
}
\description{
Summarize date distribution by year
}
\examples{
# summary_date(sale, "dot", 2010:2019)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()},
\code{\link{plot_lic_types}()},
\code{\link{summary_lic_types}()},
\code{\link{summary_year_dot}()}
}
\concept{functions to summarize license data}
