% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_duration}
\alias{summary_duration}
\title{Summarize distribution of end_date - start_date by year}
\usage{
summary_duration(
  df,
  yrs,
  start = "start_date",
  end = "end_date",
  dot = "dot",
  samp_size = 1e+05,
  facet = function() facet_wrap(~.data$year)
)
}
\arguments{
\item{df}{data frame with date variable}

\item{yrs}{years of interest}

\item{start}{variable name for start date}

\item{end}{variable name for end date}

\item{dot}{variable name for transaction date}

\item{samp_size}{sample size to use for efficiency. If NULL, no sample will
be used.}

\item{facet}{facetting function to apply to the plot}
}
\description{
Summarize distribution of end_date - start_date by year
}
\examples{
# summary_duration(sale, 2010:2019)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()},
\code{\link{plot_lic_types}()},
\code{\link{summary_date}()},
\code{\link{summary_lic_types}()},
\code{\link{summary_year_dot}()}
}
\concept{functions to summarize license data}
