% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_lic_types}
\alias{summary_lic_types}
\title{Count license type customers by year}
\usage{
summary_lic_types(df, types = c("hunt", "fish", "combo"))
}
\arguments{
\item{df}{data frame with sale data and license type info}

\item{types}{types to include (based on "type" variable)}
}
\description{
Count license type customers by year
}
\examples{
# sale <- left_join(sale, lic, by = "lic_id")
# summary_lic_types(sale)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()},
\code{\link{plot_lic_types}()},
\code{\link{summary_year_dot}()}
}
\concept{functions to summarize license data}
