% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_year_dot}
\alias{summary_year_dot}
\title{Summarize difference between "year" and year(dot) in sales}
\usage{
summary_year_dot(df, lastyr = 2019)
}
\arguments{
\item{df}{input sales dataset with "year" and "dot" variables}

\item{lastyr}{for descending sort of the output table}
}
\description{
Agency provide year variables don't always correspond to calendar year of
sales. The resulting table tabulates the absolute difference by df$year.
}
\examples{
# summary_year_dot(sale)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()},
\code{\link{plot_lic_types}()},
\code{\link{summary_lic_types}()}
}
\concept{functions to summarize license data}
