% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-check.R
\name{variable_allowed_dates}
\alias{variable_allowed_dates}
\title{Get a set of all dates that satisfy given conditions}
\usage{
variable_allowed_dates(
  year1,
  valid_years = year1:substr(Sys.Date(), 1, 4),
  valid_months = c(paste0("0", 1:9), 10:12),
  valid_days = c(paste0("0", 1:9), 10:31)
)
}
\arguments{
\item{year1}{first allowed year}

\item{valid_years}{set of allowed years}

\item{valid_months}{set of allowed months}

\item{valid_days}{set of allwed days}
}
\description{
This is intended to be called from \code{\link{data_check_standard}}
for checkiing date variables. It returns a vector of dates in a "yyyy-mm-dd"
format.
}
\seealso{
Other functions to check data format: 
\code{\link{data_check_standard}()}
}
\concept{functions to check data format}
