% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_demo}
\alias{plot_demo}
\title{Make distribution plot(s) for demographic variable(s)}
\usage{
plot_demo(df, vars, trend = FALSE)
}
\arguments{
\item{df}{data frame with demographic variable(s)}

\item{vars}{name of demographic variable(s)}

\item{trend}{If TRUE, use stacked bar plots to trend across years}
}
\description{
Make distribution plot(s) for demographic variable(s)
}
\examples{
# plot_demo(cust, c("sex", "birth_year", "cust_res"))
# hunt <- left_join(hunt, cust, by = "cust_id")
# plot_demo(hunt, c("sex", "cust_res"), trend = TRUE)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_lic_types}()},
\code{\link{summary_lic_types}()}
}
\concept{functions to summarize license data}
