% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_lic_types}
\alias{plot_lic_types}
\title{Make a heatmap of license type customers by year}
\usage{
plot_lic_types(df, types = c("hunt", "fish", "combo"), truncate = 30)
}
\arguments{
\item{df}{data frame with sale data and license type info}

\item{types}{types to include (based on "type" variable)}

\item{truncate}{Truncation point for "description" variable. If NULL, no
truncation will be performed.}
}
\description{
Make a heatmap of license type customers by year
}
\examples{
# sale <- left_join(sale, lic, by = "lic_id")
# plot_lic_types(sale)
}
\seealso{
Other functions to summarize license data: 
\code{\link{plot_demo}()}
}
\concept{functions to summarize license data}
